LayoutTests/platform/mac-catalina/fast/text/international/system-language/navigator-language/navigator-language-ru-expected.txt

FAIL navigator.language should be ru-RU. Was ru.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
